backend/lambdas/requirements.txt
boto3==1.26.137
requests==2.31.0
google-analytics-data==0.16.1
google-auth==2.17.3
python-dotenv==1.0.0 